% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_layers_leaflet.R
\name{add_layers_leaflet}
\alias{add_layers_leaflet}
\title{Add layer to \code{Leaflet} web map}
\usage{
add_layers_leaflet(
  map_object,
  map_active_df,
  map_var,
  map_colour,
  opacity,
  map_line_width,
  map_line_colour,
  waiter
)
}
\arguments{
\item{in_df}{A \code{sf} data frame object that is displayed on the leaflet
web map.}

\item{layer_id}{The layerId of the feature selected by the user.}

\item{label_vars}{The variables from \code{in_df} selected to be displayed in
the popup.}
}
\value{
\code{popup_text} a html popup label to be displayed on the
  \code{Leaflet} web map.
}
\description{
\code{add_layers_leaflet} adds a user selected spatial layer to a Leaflet web
map. \code{sf} \code{POINT}, \code{LINESTRING}, \code{MULTILINESTRING}, \code{POLYGON}, \code{MULITPOLYGON}
geometries can be added to the web map.
}
