% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_layers_leaflet_no_zoom.R
\name{add_layers_leaflet_no_zoom}
\alias{add_layers_leaflet_no_zoom}
\title{Add layer to \code{Leaflet} web map - no zoom}
\usage{
add_layers_leaflet_no_zoom(
  map_object,
  map_active_df,
  map_var,
  map_colour,
  opacity,
  map_line_width,
  map_line_colour,
  waiter
)
}
\arguments{
\item{map_object}{single-element character vector of the ID of the Leaflet map object.}

\item{map_active_df}{the spatial data frame containing features to draw on the Leaflet map object.}

\item{map_var}{the column in \code{map_active_df} to map fill / marker colours to data values when drawing the features.}

\item{map_colour}{fill colour palette - RColorBrewer palettes such as \code{"YlOrRd"}.}

\item{opacity}{feature opacity - numeric value between 0 and 1.}

\item{map_line_width}{stroke width for line and polygon features.}

\item{map_line_colour}{stroke colour for line and polygon features.}

\item{waiter}{waiter object to display while map is rendering.}
}
\value{
Leaflet proxy object.
}
\description{
\code{add_layers_leaflet} adds a user selected spatial layer to a Leaflet web
map. \code{sf} \code{POINT}, \code{LINESTRING}, \code{MULTILINESTRING},
\code{POLYGON}, \code{MULITPOLYGON} geometries can be added to the web map.
Calling this function does not change the zoom or extent of the Leaflet map
to fit the contents drawn on the map. Used for keeping the zoom extent the
same but updating colour and opacity.
}
