% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_spatial_join_tables.R
\name{db_spatial_join_tables}
\alias{db_spatial_join_tables}
\title{Spatial join of two layers.}
\usage{
db_spatial_join_tables(
  con,
  left_names,
  right_names,
  left_table_name,
  right_table_name
)
}
\arguments{
\item{con}{connection object to PostGIS database}

\item{left_names}{columns in left table to retain}

\item{right_names}{columns in right table to retains}

\item{left_table_name}{name of left table created in PostGIS}

\item{right_table_name}{name of right table created in PostGIS}
}
\description{
Spatial join of two layers based on largest intersection using PostGIS.
}
