% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_edit_data_frame.R
\name{edit_data_frame}
\alias{edit_data_frame}
\title{Apply edits from user to layers in a GeoPackage}
\usage{
edit_data_frame(tmp_edits, df_to_edit, df_to_edit_not_sf, layer)
}
\arguments{
\item{tmp_edits}{Data frame recording user supplied edits (derived from
editing DataTables object in the UI).}

\item{df_to_edit}{Data frame or spatial data frame (sf) representing the layer in
GeoPackage to apply user edits to.}

\item{df_to_edit_not_sf}{Data frame representing the layer in GeoPackage with the sf
geometry object dropped.}

\item{layer}{Name of layer that edits are being applied to.}
}
\value{
A two-element list. The first element is the data frame with edits
 applied to it. The second element is a log recording the status of attempts
 to apply user edits to the underlying data frame object.
}
\description{
\code{edit_data_frame} applies user edits to a DataTables object
in a browser to the underlying (spatial) data frame.
}
