% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_gcs_object.R
\name{get_gcs_object}
\alias{get_gcs_object}
\title{Get GeoPackages from Google Cloud Storage bucket}
\usage{
get_gcs_object(token, bucket, object)
}
\arguments{
\item{token}{an OAuth2.0 token which allows a user to make authenticated HTTP
requests to the Google Cloud Storage API.}

\item{bucket}{name of Google Cloud Storage bucket.}

\item{object}{name of GeoPackage to get.}
}
\value{
two element list. First element is the path to where the GeoPackage
  has been written or character vector with an error message. The second
  element is the file name to display to users when selecting the GeoPackage.
}
\description{
Makes a request to a user specified Google Cloud Storage bucket to get object
(GeoPackage) in the bucket.
}
