% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qfieldcloud.R
\name{get_qfieldcloud_gpkg}
\alias{get_qfieldcloud_gpkg}
\title{Get QFieldCloud selected GeoPackage in project}
\usage{
get_qfieldcloud_gpkg(token, endpoint, project_id, filename)
}
\arguments{
\item{token}{session token}

\item{endpoint}{QFieldCloud app url (omit https:// and trailing /)}

\item{project_id}{project id corresponding to project to download file from}

\item{filename}{name of GeoPacakge to download}
}
\value{
list with path to downloaded GeoPackage and filename
}
\description{
Get QFieldCloud selected GeoPackage in project
}
