% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qfieldcloud.R
\name{get_qfieldcloud_projects}
\alias{get_qfieldcloud_projects}
\title{Get QFieldCloud projects}
\usage{
get_qfieldcloud_projects(token, endpoint)
}
\arguments{
\item{token}{session token}

\item{endpoint}{QFieldCloud app url (omit https:// and trailing /)}
}
\value{
data.frame with two columns storing project names and project ids
}
\description{
Get QFieldCloud projects
}
