% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_group_by_summarise.R
\name{group_by_summarise}
\alias{group_by_summarise}
\title{Group by and summarise}
\usage{
group_by_summarise(in_df, group_var, summ_var)
}
\arguments{
\item{in_df}{A table / layer (spatial or non-spatial) to summarise.}

\item{group_var}{A character vector of variable names to group by - uses
\code{\link[dplyr]{group_by}}.}

\item{summ_var}{A character vector of variable names to summarise - uses
\code{\link[dplyr]{summarise}}. For all numeric variables the mean and sum
are computed. For numeric and non-numeric variables the count of
observations within each group is returned.}
}
\value{
A summary table as a data frame (\code{s_df}) object.
}
\description{
Performs group by and summarise operations on user selected grouping
variables and variables to compute summary statistics within a group.
}
