% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_list_gcs_bucket_objects.R
\name{list_gcs_bucket_objects}
\alias{list_gcs_bucket_objects}
\title{List GeoPackages in Google Cloud Storage bucket}
\usage{
list_gcs_bucket_objects(token, bucket)
}
\arguments{
\item{token}{a \href{https://httr.r-lib.org/reference/oauth2.0_token.html}{httr OAuth2.0 token} which allows a user to make authenticated HTTP
requests to the Google Cloud Storage API.}

\item{bucket}{name of Google Cloud Storage bucket to list objects in.}
}
\value{
items - a character vector of the names of GeoPacakges in the Google
  Cloud Storage bucket.
}
\description{
Makes a request to a user specified Google Cloud Storage bucket to list all
objects in the bucket. Filters the response to return a list of GeoPackages
(if any). This function uses the \href{https://httr.r-lib.org/reference/index.html}{httr} package for making authenticated
HTTP requests and processing responses.
}
