% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_list_gcs_buckets.R
\name{list_gcs_buckets}
\alias{list_gcs_buckets}
\title{List Google Cloud Storage buckets}
\usage{
list_gcs_buckets(token, project_id)
}
\arguments{
\item{token}{an OAuth2.0 token which allows a user to make authenticated HTTP
requests to the Google Cloud Storage API.}

\item{project_id}{single element character vector of Google Cloud project id.}
}
\value{
vector of bucket names.
}
\description{
Makes a request to a user specified Google Cloud project to list storage bucekts
}
