% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_list_layers.R
\name{list_layers}
\alias{list_layers}
\title{Return layers from uploaded data}
\usage{
list_layers(f_path, f_name)
}
\arguments{
\item{f_path}{A character vector of file paths for GeoPackage(s) uploaded by the user.}

\item{f_name}{A character vector of file names of GeoPackage(s) uploaded by the user.}
}
\value{
A tibble with three columns: layers = table / layer name;
  layer_disp_name = clean and informative table / layer name for select
  input; file_path = temporary file path to data; and file_type = file type.
}
\description{
\code{list_layers} returns the name and path to tables / layers from data
uploaded to the application as a tibble object. If a GeoPackage is uploaded which
contains more than one table, each table and the path to that table will be
recorded on a separate row of the returned tibble.
}
