% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qfieldcloud.R
\name{list_qfieldcloud_gpkg}
\alias{list_qfieldcloud_gpkg}
\title{Get QFieldCloud files in project (GeoPackages only)}
\usage{
list_qfieldcloud_gpkg(token, endpoint, project_id)
}
\arguments{
\item{token}{session token}

\item{endpoint}{QFieldCloud app url (omit https:// and trailing /)}

\item{project_id}{project id corresponding to project to download file from}
}
\value{
data.frame with two columns storing file names and project ids
}
\description{
Get QFieldCloud files in project (GeoPackages only)
}
