% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_barplot.R
\name{make_barplot}
\alias{make_barplot}
\title{Barplot}
\usage{
make_barplot(col_chart_df, x_lab, y_lab, axis_font_size, lab_font_size)
}
\arguments{
\item{col_chart_df}{a two column data frame. The first column specifies the
groupings. The second column specifies the summary statistic per-group.}

\item{x_lab}{a single element character vector specifying the X-axis label.}

\item{y_lab}{a single element character vector specifying the Y-axis label.}

\item{axis_font_size}{a numeric value specifying the font size for axis tick
labels.}

\item{lab_font_size}{a numeric value specifying the font size for axis
labels.}
}
\value{
a \link[ggplot2]{ggplot} object that can be passed to
  \link[shiny]{renderPlot} to display in a Shiny application.
}
\description{
Generate and style a bar plot which visualises summary statistics (mean, sum,
count) of values in a group. Designed to be used in  a Shiny application to
visualise results of \link[dplyr]{group_by} and
\link[dplyr]{summarise} operations.
}
\details{
This is a helper function to pass user defined data, aesthetic mappings, and chart styling parameters to \link[ggplot2]{ggplot} within a Shiny app.
}
