% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_histogram.R
\name{make_histogram}
\alias{make_histogram}
\title{Histogram}
\usage{
make_histogram(
  chart_active_df,
  hist_x_var,
  binwidth,
  x_lab,
  y_lab,
  axis_font_size,
  lab_font_size
)
}
\arguments{
\item{chart_active_df}{data frame containing a column whose values will be
used to generate a histogram.}

\item{hist_x_var}{a single element character vector specifying the name of
the column to use to generate a histogram.}

\item{binwidth}{numeric value specifying the width of the histogram's bins.}

\item{x_lab}{a single element character vector specifying the X-axis label.}

\item{y_lab}{a single element character vector specifying the Y-axis label.}

\item{axis_font_size}{a numeric value specifying the font size for axis tick
labels.}

\item{lab_font_size}{a numeric value specifying the font size for axis
labels.}
}
\value{
a \link[ggplot2]{ggplot} object that can be passed to
  \link[Shiny]{renderPlot} to display in a Shiny application.
}
\description{
Generate and style a histogram from a column in a (spatial -
\href{https://r-spatial.github.io/sf/index.html}{sf}) data frame. Designed to
be used in a Shiny application to visualise the distribution of continuous
variables in layers in a GeoPackage.
}
