% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scatter.R
\name{make_scatter}
\alias{make_scatter}
\title{Scatter plot}
\usage{
make_scatter(
  chart_active_df,
  scatter_x_var,
  scatter_y_var,
  point,
  x_lab,
  y_lab,
  axis_font_size,
  lab_font_size
)
}
\arguments{
\item{chart_active_df}{data frame containing a column whose values will be
used to generate a histogram.}

\item{scatter_x_var}{a single element character vector specifying the name of
the column to use to plot on the X-axis.}

\item{scatter_y_var}{a single element character vector specifying the name of
the column to use to plot on the Y-axis.}

\item{point}{point size.}

\item{x_lab}{a single element character vector specifying the X-axis label.}

\item{y_lab}{a single element character vector specifying the Y-axis label.}

\item{axis_font_size}{a numeric value specifying the font size for axis tick
labels.}

\item{lab_font_size}{a numeric value specifying the font size for axis
labels.}
}
\value{
a \link[ggplot2]{ggplot} object that can be passed to
  \link[Shiny]{renderPlot} to display in a Shiny application.
}
\description{
Generate and style a scatter plot from two numeric columns in a (spatial -
\href{https://r-spatial.github.io/sf/index.html}{sf}) data frame. Designed to
be used in a Shiny application to visualise the bivariate relationship between two variables.
}
\details{
This is a helper function to pass user defined data, aesthetic mappings, and chart styling parameters to \link[ggplot2]{ggplot} within a Shiny app.
}
