% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_make_spatial_db_table.R
\name{make_spatial_db_table}
\alias{make_spatial_db_table}
\title{Add right and left tables to PostGIS and prepare for spatial join}
\usage{
make_spatial_db_table(
  con,
  left_table,
  left_table_name,
  right_table,
  right_table_name
)
}
\arguments{
\item{con}{connection object to PostGIS database}

\item{left_table}{sf spatial data frame of left table in spatial join. Geometries for this object will be retained.}

\item{left_table_name}{name of left table created in PostGIS}

\item{right_table}{sf spatial data frame of right table in spatial join. Geometries for this object are dropped and column names are added to `left_table` based on the largest intersection.}

\item{right_table_name}{name of right table created in PostGIS}
}
\description{
Add right and left tables to PostGIS, convert to geography type, and create spatial indexes.
}
