% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_get_layers.R
\name{mod_get_layers_Server}
\alias{mod_get_layers_Server}
\title{get_layers Server}
\usage{
mod_get_layers_Server(id)
}
\arguments{
\item{session}{}
}
\value{
\code{field_layers} tibble with three columns: layers - layer name;
  layer_disp_name - clean and informative layer name for select
  input; file_path - temporary file path to data; and file_type - file type.
}
\description{
Process files (GeoPackages) uploaded by the user. For each file
  uploaded, call \code{list_layers()} to list all layers in the GeoPacakge
  and the temporary location of the GeoPackage.
}
