% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_get_layers.R
\name{mod_get_layers_UI}
\alias{mod_get_layers_UI}
\title{Upload GeoPackages to a Shiny App}
\usage{
mod_get_layers_UI(id, label, multiple, accept)
}
\arguments{
\item{id}{inputID for file upload UI component.}

\item{label}{label for file upload widget on the UI.}

\item{multiple}{logical, whether to accept multiple file uploads.}

\item{accept}{file types to accept - vector of strings indicating extensions or MIME type}
}
\description{
Generate UI component for file upload using Shiny's \code{fileInput} to allow users to upload GeoPackages to a Shiny app.
}
