% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plant_number.R
\name{plant_number}
\alias{plant_number}
\title{Compute plant number}
\usage{
plant_number(acres, plant_spacing, row_spacing)
}
\arguments{
\item{acres}{The area of the crop in acres.}

\item{plant_spacing}{The plant spacing of between plants in a row in metres.}

\item{row_spacing}{The row spacing between rows of plants in metres.}
}
\value{
The number of plants in the field.
}
\description{
Compute plant number
}
