% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qfieldcloud.R
\name{qfieldcloud_login}
\alias{qfieldcloud_login}
\title{QFieldCloud login}
\usage{
qfieldcloud_login(username, password, endpoint)
}
\arguments{
\item{username}{QFieldCloud email}

\item{password}{QFieldCloud password}

\item{endpoint}{QFieldCloud app url (omit https:// and trailing /)}
}
\value{
session token
}
\description{
QFieldCloud login
}
