% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_tables.R
\name{read_tables}
\alias{read_tables}
\title{Import a table - either spatial a table (i.e. a table in a GeoPackage) or csv file.}
\usage{
read_tables(uploads, lyr)
}
\arguments{
\item{uploads}{Data frame generated from \code{list_layers} indicating names, file types, and paths to files uploaded tables.}

\item{lyr}{User selected layer - this is used for selecting the relevant layer from a GeoPackage with many tables.}
}
\value{
Table - data frame or spatial table (of class \code{sf}) of the user selected layer.
}
\description{
Import a table - either spatial a table (i.e. a table in a GeoPackage) or csv file.
}
