% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_layers_leaflet.R
\name{add_layers_leaflet}
\alias{add_layers_leaflet}
\title{Add layer to a Leaflet web map}
\usage{
add_layers_leaflet(
  map_object,
  map_active_df,
  map_var,
  map_colour,
  opacity,
  map_line_width,
  map_line_colour,
  waiter
)
}
\arguments{
\item{map_object}{single-element character vector of the ID of the Leaflet map object.}

\item{map_active_df}{the spatial data frame containing features to draw on the Leaflet map object.}

\item{map_var}{the column in \code{map_active_df} to map fill / marker colours to data values when drawing the features.}

\item{map_colour}{fill colour palette - RColorBrewer palettes such as \code{"YlOrRd"}.}

\item{opacity}{feature opacity - numeric value between 0 and 1.}

\item{map_line_width}{stroke width for line and polygon features.}

\item{map_line_colour}{stroke colour for line and polygon features.}

\item{waiter}{waiter object to display while map is rendering.}
}
\value{
Leaflet proxy object.
}
\description{
\code{add_layers_leaflet} adds a user selected spatial layer to a Leaflet web
map. sf data frames with \code{POINT}, \code{LINESTRING}, \code{MULTILINESTRING}, \code{POLYGON}, \code{MULITPOLYGON}
geometries can be added to the web map.
}
