% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_layers_leaflet_no_zoom.R
\name{add_layers_leaflet_no_zoom}
\alias{add_layers_leaflet_no_zoom}
\title{Add layer to \code{Leaflet} web map - no zoom}
\usage{
add_layers_leaflet_no_zoom(
  map_object,
  map_active_df,
  map_var,
  map_colour,
  opacity,
  map_line_width,
  map_line_colour,
  waiter
)
}
\arguments{
\item{in_df}{A \code{sf} data frame object that is displayed on the leaflet
web map.}

\item{layer_id}{The layerId of the feature selected by the user.}

\item{label_vars}{The variables from \code{in_df} selected to be displayed in
the popup.}
}
\value{
\code{popup_text} a html popup label to be displayed on the
  \code{Leaflet} web map.
}
\description{
\code{add_layers_leaflet} adds a user selected spatial layer to a Leaflet web
map. \code{sf} \code{POINT}, \code{LINESTRING}, \code{MULTILINESTRING},
\code{POLYGON}, \code{MULITPOLYGON} geometries can be added to the web map.
Calling this function does not change the zoom or extent of the Leaflet map
to fit the contents drawn on the map. Used for keeping the zoom extent the
same but updating colour and opacity.
}
