% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_popups.R
\name{add_popups}
\alias{add_popups}
\title{Add popup label when feature is clicked on a \code{Leaflet} web map}
\usage{
add_popups(in_df, layer_id, label_vars)
}
\arguments{
\item{in_df}{A \code{sf} data frame object that is displayed on the leaflet web
map.}

\item{layer_id}{The layerId of the feature selected by the user.}

\item{label_vars}{The variables from \code{in_df} selected to be displayed in
the popup.}
}
\value{
\code{popup_text} a html popup label to be displayed on the \code{Leaflet}
web map.
}
\description{
\code{add_popups} creates a popup label containing values of selected
variables for a feature selected through an on-click event on a \code{Leaflet}
web map.
}
