% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_filter_rows.R
\name{filter_rows}
\alias{filter_rows}
\title{Filter rows from a (spatial) data frame based on a condition}
\usage{
filter_rows(filter_df, filter_conditions)
}
\arguments{
\item{filter_df}{data frame or spatial (sf) data frame to apply condition to
and filter rows from.}

\item{filter_conditions}{conditions used to select rows to filter from
\code{filter_df}. Conditions should be specified in R or dplyr syntax (e.g.
\code{column_1 == "xyz"} will filter all rows from \code{filter_df} where
\code{column_1} has a value of \code{"xyz"}).}
}
\value{
a data frame or spatial (sf) data frame containing the records
  filtered from \code{filter_df}.
}
\description{
Designed to be used within Shiny applications. Allows a user to specify
conditions (using R or dplyr syntax) to filter rows from a data frame or
spatial (sf) data frame. This function uses dplyr's \code{filter()}.
}
