% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_get_layers.R
\name{mod_get_layers_UI}
\alias{mod_get_layers_UI}
\title{get_layers UI}
\usage{
mod_get_layers_UI(id, label, multiple, accept)
}
\arguments{
\item{id}{inputID for file upload UI component.}

\item{label}{label for file upload widget on the UI.}

\item{multiple}{logical, whether to accept multiple file uploads.}

\item{accept}{file types to accept - vector of strings indicating extensions or MIME type}
}
\value{

}
\description{
Generate UI for file upload.
}
