% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_tables.R
\name{read_tables}
\alias{read_tables}
\title{Read tabular data - either spatial layer (i.e. a table in a GeoPackage) or csv file.}
\usage{
read_tables(uploads, lyr)
}
\arguments{
\item{uploads}{Data frame generated from \code{list_layers} indicating names, file types, and paths to uploaded layers.}

\item{lyr}{User selected layer - this is used for selecting the relevant layer from a GeoPackage with many layers.}
}
\value{
a (spatial) data frame of the selected layer.
}
\description{
Read tabular data - either spatial layer (i.e. a table in a GeoPackage) or csv file.
}
