% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spatial_join_tables.R
\name{spatial_join_tables}
\alias{spatial_join_tables}
\title{Spatial join of two layers}
\usage{
spatial_join_tables(f_left, f_right, j_type)
}
\arguments{
\item{f_left}{left table in spatial joins - a spatial
(\href{https://r-spatial.github.io/sf/index.html}{sf}) data frame.}

\item{f_right}{Right table in spatial joins - a spatial
(\href{https://r-spatial.github.io/sf/index.html}{sf}) data frame.}

\item{j_type}{A character specifying the type of join to perform. Options for
spatial joins include inner and left joins.}
}
\value{
\code{j_df} A layer of class - a spatial
  (\href{https://r-spatial.github.io/sf/index.html}{sf}) data frame (i.e.
  spatial features and attribute columns) returned by the spatial join.
}
\description{
Perform a spatial inner join or left join operation on two layers based on
largest intersection between spatial features. Spatial joins are performed
using \href{https://r-spatial.github.io/sf/index.html}{sf}'s
\link[sf]{st_join}.
}
\details{
Geometries of both spatial tables are checked for validity (using
\href{https://r-spatial.github.io/sf/index.html}{sf}'s \code{st_is_valid})
and that they are not empty (using
\href{https://r-spatial.github.io/sf/index.html}{sf}'s \code{st_is_empty}).
Invalid geometries are dropped before performing spatial joins.
}
