% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_write_tables.R
\name{write_tables}
\alias{write_tables}
\title{Write a layer to a GeoPackage file}
\usage{
write_tables(df, gpkg_dir, lyr)
}
\arguments{
\item{df}{Data frame or spatial data frame (sf) to write to GeoPackage.}

\item{gpkg_dir}{Data frame generated from \code{list_layers} indicating names, file types, and paths to files uploaded tables.}

\item{lyr}{User selected layer - this is used for selecting the relevant layer from a GeoPackage with many tables.}
}
\description{
Write a layer to a GeoPackage file
}
